Ensure numpy array sent to segmenter is of right dtype

+++ src/iterseg/segmentation.py
@@ -862,7 +862,8 @@ def segmentation_loop(
     ndim = data.ndim
     if ndim == 3:
         output = segment_single_volume(
-                np.asarray(data), chunk_size, config, margin,
+                np.asarray(data).astype(np.float32),
+                chunk_size, config, margin,
                 processing_function
                 )
         output_labels[...] = output
